Move Instruction SIB bug fix
Fixes bug introduced by commit 385077a for operand 89 where any
ModRM rm field of 4 or greater would be considered a SIB address.
The intended behaviour was for the field to be treated as a SIB
address only if the rm field equals 4. We have introduced a new
unit test for this bug as well.