Fixed SIB displacement
SIB byte with base of 5 is a special case. When the ModRM mod-bit
is set to 0, we have no SIB base, and a 32-bit displacement. In
all other cases, this should act like a normal SIB address. Previous
code was treating SIB base 5 with mod 1 or 2 as no displacement.
Fixed with tests.