Modified underlying stack to work more like x86.
The previous stack implementation was simply an array that held
values as they were pushed or popped. The stack pointer started
at zero and increased as items were added. The x86 stack does
not work this way, this commit brings the stack more in line
with the behaviour of the x86 stack. Unit tests amended.